requirements.txt
numpy~=1.21.2
pyyaml~=6.0
matplotlib~=3.5.1
pillow~=9.0.1
pynput~=1.7.6
cvxpy~=1.2.0
pathlib~=1.0.1
imageio~=2.16.1
setuptools~=58.0.4
opencv-python~=4.1.2.30